Updating to released hub-common 22.0.0
