+++ build.gradle
@@ -40,7 +40,7 @@ repositories {
 }
 
 dependencies {
-    compile 'com.blackducksoftware.integration:hub-common:22.0.0-SNAPSHOT'
+    compile 'com.blackducksoftware.integration:hub-common:22.0.0'
     compile 'org.springframework.boot:spring-boot-starter'
     compile 'org.springframework.boot:spring-boot-starter-data-jpa'
     compile 'org.springframework.boot:spring-boot-starter-batch'
